elasticsearch6: refactor plugins
The plugins command didn't actually work with elasticsearch6 so
I have added a new derivation for elasticsearch6 plugins. This
also means there won't be problems with versions of plugins that
only work for ES5 and not ES6.

Unfortunately it is down to the user to use the elasticsearch6Plugins
packages when setting up an elasticsearch server but I couldn't
think of a way around this.

Currently there is only 1 plugin, the discovery ec2 plugin since
this is the one I am using.